requirements.txt
actionlib==1.14.0
annotated-types==0.7.0
bondpy==1.8.6
catkin==0.8.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cv-bridge==1.16.2
cycler==0.12.1
dill==0.3.9
docker-pycreds==0.4.0
dynamic-reconfigure==1.7.3
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
kmeans-pytorch==0.3
MarkupSafe==3.0.2
matplotlib==3.10.0
message-filters==1.17.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytorch_revgrad==0.2.0
pytz==2025.1
PyYAML==6.0.2
qt-gui==0.4.2
requests==2.32.3
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt_bag==0.5.1
rqt_gui==0.5.3
rqt_gui_py==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sensor-msgs==1.13.1
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smclib==1.8.6
smmap==5.0.2
sympy==1.13.1
tf==1.13.2
tf-conversions==1.13.2
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.5.0
topic-tools==1.17.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.3.0
wandb==0.19.7
